deID/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Faker==22.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-cov==4.1.0
